[Chore #163846710] Move erorrhandlers to module
