+++ app/__init__.py
@@ -6,6 +6,7 @@ Creation of application factory
 from flask import Flask, jsonify
 
 from instance.config import config_environment
+from app.api.util.error_handlers import bad_request, internal_server_error, url_not_found, method_not_allowed
 from app.api.controller.party_api import parties as parties_bp
 from app.api.controller.office_api import offices as offices_bp
 
@@ -22,46 +23,11 @@ def create_app(config_name):
     app = Flask(__name__, instance_relative_config=True)
     app.config.from_object(config_environment[config_name])
 
-    # Register app error handlers
-    @app.errorhandler(405)
-    def method_not_allowed(e):
-        """
-        Json response for Method Not Allowed.
-        """
-        return jsonify({
-            "status": 405,
-            "error": "Method Not Allowed on Url."
-        }), 405
-
-    @app.errorhandler(500)
-    def internal_server_error(e):
-        """
-        custom json response when server hits a 500
-        """
-        return jsonify({
-            "status": 500,
-            "error": "Our developers are looking into the issue."
-        }), 500
-
-    @app.errorhandler(400)
-    def bad_request(e):
-        """
-        Custom json response for bad request at app level
-        """
-        return jsonify({
-            "status": 400,
-            "error": "Check your request format."
-        }), 400
-
-    @app.errorhandler(404)
-    def url_not_found(e):
-        """
-        Custom json response for bad request at app level
-        """
-        return jsonify({
-            "status": 404,
-            "error": "Check url and try again."
-        }), 404
+    # register error handlers
+    app.register_error_handler(400, bad_request)
+    app.register_error_handler(404, url_not_found)
+    app.register_error_handler(500, internal_server_error)
+    app.register_error_handler(405, method_not_allowed)
 
     # register blueprints to app
     app.register_blueprint(parties_bp, url_prefix='/api/v1')
